Extract `prepare_label_group` function from `write.py`
