+++ .gitignore
@@ -1,4 +1,5 @@
 .pytest_cache/
+tree.sh
 .coverage*
 __pycache__
 *.cpython-*.pyc
