+++ fractal_tasks_core/tasks/cellpose_segmentation.py
@@ -36,9 +36,9 @@ from fractal_tasks_core.ome_zarr.channels import ChannelNotFoundError
 from fractal_tasks_core.ome_zarr.channels import get_channel_from_image_zarr
 from fractal_tasks_core.ome_zarr.channels import OmeroChannel
 from fractal_tasks_core.ome_zarr.input_models import Channel
+from fractal_tasks_core.ome_zarr.label_group import prepare_label_group
 from fractal_tasks_core.ome_zarr.masked_loading import masked_loading_wrapper
 from fractal_tasks_core.ome_zarr.pyramids import build_pyramid
-from fractal_tasks_core.ome_zarr.write import prepare_label_group
 from fractal_tasks_core.ome_zarr.zattrs_utils import rescale_datasets
 from fractal_tasks_core.roi import (
     array_to_bounding_box_table,
